requirements.txt
pytest==7.4.3
requests==2.31.0
jsonschema==4.20.0
selene~=2.0.0rc4
allure-pytest==2.9.45
